+++ ts/features/wallet/component/features/BasePaymentFeatureListItem.tsx
@@ -1,0 +1,46 @@
+import { View } from "native-base";
+import * as React from "react";
+import { StyleSheet } from "react-native";
+import { H4 } from "../../../../components/core/typography/H4";
+import { H5 } from "../../../../components/core/typography/H5";
+import { IOStyles } from "../../../../components/core/variables/IOStyles";
+import { WithTestID } from "../../../../types/WithTestID";
+
+type Props = WithTestID<{
+  title: string;
+  description: string;
+  rightElement: React.ReactNode;
+}>;
+
+const styles = StyleSheet.create({
+  row: {
+    paddingVertical: 16,
+    flexDirection: "row",
+    justifyContent: "space-between"
+  },
+  left: {
+    ...IOStyles.flex,
+    paddingRight: 8
+  }
+});
+
+/**
+ * The base layout for the Payment feature list items
+ * @param props
+ * @constructor
+ */
+export const BasePaymentFeatureListItem = (
+  props: Props
+): React.ReactElement => (
+  <View style={styles.row} testID={props.testID}>
+    <View style={styles.left}>
+      <H4 weight={"SemiBold"} color={"bluegreyDark"}>
+        {props.title}
+      </H4>
+      <H5 weight={"Regular"} color={"bluegrey"}>
+        {props.description}
+      </H5>
+    </View>
+    {props.rightElement}
+  </View>
+);
