+++ ts/features/wallet/component/features/PaymentMethodSettings.tsx
@@ -3,13 +3,13 @@ import * as React from "react";
 import { StyleSheet } from "react-native";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import { H3 } from "../../../components/core/typography/H3";
-import { IOColors } from "../../../components/core/variables/IOColors";
-import IconFont from "../../../components/ui/IconFont";
-import I18n from "../../../i18n";
-import { GlobalState } from "../../../store/reducers/types";
-import { PaymentMethod } from "../../../types/pagopa";
-import PagoPaPaymentCapability from "./PagoPaPaymentCapability";
+import { H3 } from "../../../../components/core/typography/H3";
+import { IOColors } from "../../../../components/core/variables/IOColors";
+import IconFont from "../../../../components/ui/IconFont";
+import I18n from "../../../../i18n";
+import { GlobalState } from "../../../../store/reducers/types";
+import { PaymentMethod } from "../../../../types/pagopa";
+import PagoPaPaymentCapability from "../PagoPaPaymentCapability";
 
 type OwnProps = { paymentMethod: PaymentMethod };
